feat: implement useHttpExampleFetch and useHttpExampleState for data fetching and state management
